+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog of nens-auth-client
 ===================================================
 
 
-0.6 (unreleased)
+0.6 (2021-01-11)
 ----------------
 
 - Made all 403 error messages configurable. Some errors from the accept_invitation